More cohort action components to a separate file
